requirements.txt
certifi==2023.11.17
charset-normalizer==3.3.2
idna==3.6
pyTelegramBotAPI==4.15.2
python-dateutil==2.8.2
python-dotenv==1.0.0
python-telegram-bot-calendar==1.0.5
requests==2.31.0
six==1.16.0
telebot==0.0.5
urllib3==1.26.15
